pyproject.toml
[tool.poetry]
name = "smart_control"
version = "0.1.0"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10.12,<3.12"
dm-reverb = {extras = ["tensorflow"], version = "^0.14.0"}
pytz = "^2024.1"
gin-config = "^0.5.0"
matplotlib = "^3.9.2"
pandas = "^2.2.2"
bidict = "^0.23.1"
protobuf = "3.20.3"
pint = "^0.24.3"
holidays = "^0.56"
google3 = "^1.0.0"
opencv-python = "^4.10.0.84"
scipy = "^1.14.1"
mediapy = "^1.2.2"
seaborn = "^0.13.2"
tf-agents = "0.18.0"
scikit-learn = "^1.5.1"
ipykernel = "^6.29.5"
typing-extensions = "^4.12.2"
ipython = "^8.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
